+++ chartlets.js/packages/lib/src/plugins/mui/index.ts
@@ -4,6 +4,7 @@ import { Button } from "./Button";
 import { Checkbox } from "./Checkbox";
 import { CircularProgress } from "./CircularProgress";
 import { IconButton } from "./IconButton";
+import { RadioGroup } from "./RadioGroup";
 import { Select } from "./Select";
 import { Switch } from "./Switch";
 import { Tabs } from "./Tabs";
@@ -17,6 +18,7 @@ export default function mui(): Plugin {
       ["Checkbox", Checkbox],
       ["CircularProgress", CircularProgress],
       ["IconButton", IconButton],
+      ["RadioGroup", RadioGroup],
       ["Select", Select],
       ["Switch", Switch],
       ["Tabs", Tabs],