+++ chartlets.py/chartlets/components/button.py
@@ -28,6 +28,10 @@ class Button(Component):
     One "contained" | "outlined" | "text". Defaults to "text".
     """
 
+    tooltip: str | None = None
+    """Tooltip title. Optional."""
+
+
 
 @dataclass(frozen=True)
 class IconButton(Component):
@@ -50,3 +54,7 @@ class IconButton(Component):
     """The button variant. 
     One "contained" | "outlined" | "text". Defaults to "text".
     """
+
+    tooltip: str | None = None
+    """Tooltip title. Optional."""
+